% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage Dataset}
\format{
A data frame with `15` rows and `2` variables:
\describe{
  \item{DISTANCE}{(numeric) Distance from the nearest fire station (in miles).}
  \item{DAMAGE}{(numeric) Property damage due to fire (in dollars).}
}
}
\source{
Simulated data for educational purposes.
}
\usage{
data(fire)
}
\description{
This dataset contains information about fire damage incidents, specifically examining the relationship
between the distance of the nearest fire station and the resulting property damage.
}
\examples{
data(fire)
head(fire)
}
\keyword{datasets}
