% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{getmode}
\alias{getmode}
\title{Calculate the Mode of a Vector}
\usage{
getmode(v)
}
\arguments{
\item{v}{A numeric or character vector.}
}
\value{
The mode (most frequently occurring value) of the vector.
}
\description{
Calculate the Mode of a Vector
}
\examples{
getmode(c(1, 2, 2, 3, 3, 3, 4)) # Returns 3
}
