% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Binomial Simulation Function}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{Integer. The number of iterations/simulations to run. Default is 100.}

\item{n}{Integer. The number of trials in each iteration. Default is 10.}

\item{p}{Numeric. The probability of success in each trial. Default is 0.5.}
}
\value{
A table of proportions showing the frequency of each number of successes across all iterations as well as a bar plot.
}
\description{
Binomial Simulation Function
}
\examples{
mybin()
mybin(iter = 1000, n = 10, p = 0.7)
}
