% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{Central Limit Theorem Simulation}
\usage{
myclt(n, iter, a = 0, b = 5)
}
\arguments{
\item{n}{Integer. The number of uniform random variables to sum in each iteration.}

\item{iter}{Integer. The number of iterations (samples) to generate.}

\item{a}{Numeric. The lower bound of the uniform distribution. Default is 0.}

\item{b}{Numeric. The upper bound of the uniform distribution. Default is 5.}
}
\value{
A numeric vector containing the sums of the sampled uniform random variables. Additionally, a histogram is plotted showing the distribution of the sums over all iterations, with an overlaid normal density curve.
}
\description{
Central Limit Theorem Simulation
}
\examples{
myclt(n=50, iter=10000, a=5, b=10)

}
