% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myhyper.R
\name{myhyper}
\alias{myhyper}
\title{Hypergeometric Simulation}
\usage{
myhyper(iter = 100, N = 20, r = 12, n = 5, plot = TRUE)
}
\arguments{
\item{iter}{Integer. Number of iterations.}

\item{N}{Integer. Population size.}

\item{r}{Integer. Number of successes in the population.}

\item{n}{Integer. Number of draws.}

\item{plot}{Logical. Whether to plot the results. Default is TRUE.}
}
\value{
A table of proportions representing the frequency of the number of successes across all iterations.
}
\description{
Hypergeometric Simulation
}
\examples{
#' # Run the simulation with default parameters
myhyper()

# Run the simulation with custom parameters and no plot
myhyper(iter = 500, N = 50, r = 20, n = 10, plot = FALSE)

# Run the simulation with fewer iterations and a smaller population
myhyper(iter = 50, N = 10, r = 4, n = 3)
}
