% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mynbin.R
\name{mynbin}
\alias{mynbin}
\title{Negative Binomial Mass Function}
\usage{
mynbin(y, r, p)
}
\arguments{
\item{y}{Integer. Total number of trials (must be >= `r`).}

\item{r}{Integer. Number of successes.}

\item{p}{Numeric. Probability of success.}
}
\value{
The probability of observing `r` successes in `y` trials, given a probability of success `p`.
}
\description{
Negative Binomial Mass Function
}
\examples{
mynbin(10, 3, 0.4) # Compute P(Y = 10) for a negative binomial with r = 3, p = 0.4
}
