% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Plot Normal Distribution w/Shaded Area}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{Numeric. Mean of the normal distribution.}

\item{sigma}{Numeric. Standard deviation of the normal distribution.}

\item{a}{Numeric. Value to calculate the probability P(X ≤ a).}
}
\value{
List with the mean, standard deviation, value a, and probability P(X ≤ a).
}
\description{
Plot Normal Distribution w/Shaded Area
}
\examples{
myncurve(mu = 0, sigma = 1, a = 1.5)
}
