% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myNRML.R
\name{myNRML}
\alias{myNRML}
\title{Newton-Raphson Method for Maximizing Log-Likelihood}
\usage{
myNRML(x0, delta = 0.001, llik, xrange, parameter = "param")
}
\arguments{
\item{x0}{Numeric. Initial guess for the parameter.}

\item{delta}{Numeric. Step size for numerical differentiation. Default is 0.001.}

\item{llik}{Function. A function representing the log-likelihood to be maximized.}

\item{xrange}{Numeric vector of length 2. The x-axis range for plotting.}

\item{parameter}{Character. Name of the parameter being estimated (for axis labeling). Default is "param".}
}
\value{
A list with two components:
\describe{
  \item{x}{A vector of parameter values at each iteration.}
  \item{y}{A vector of first derivative values at each iteration.}
}
}
\description{
Newton-Raphson Method for Maximizing Log-Likelihood
}
\examples{
x=c(4,6,7,6,5)
myNRML(x0=.1,delta=.0001,llik=function(x)log(dpois(4,x))+log(dpois(6,x))+log(dpois(7,x))+log(dpois(6,x))+log(dpois(5,x)),xrange=c(4,7),parameter="lambda")
}
