% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysq.R
\name{mysq}
\alias{mysq}
\title{Mysquare Function}
\usage{
mysq(x)
}
\arguments{
\item{x}{A quantitative vector}
}
\value{
A quantitative vector with squared components
}
\description{
Mysquare Function
}
\examples{
mysq(1:10)
}
