% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{ntickets}
\alias{ntickets}
\title{Compute the Number of Tickets to Sell Based on Binomial and Normal Approximation}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Integer. The number of available seats or tickets.}

\item{gamma}{Numeric. The risk threshold (e.g., probability of overselling).}

\item{p}{Numeric. The probability of a ticket holder actually attending.}
}
\value{
A list containing:
\item{N}{The original number of tickets available.}
\item{p}{The probability of a ticket being used.}
\item{gamma}{The risk threshold.}
\item{nd}{The number of tickets calculated using the binomial distribution approach.}
\item{nc}{The number of tickets calculated using the normal approximation approach.}
}
\description{
Compute the Number of Tickets to Sell Based on Binomial and Normal Approximation
}
\examples{
ntickets(400, 0.02, 0.95)
}
