% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.r
\name{pareto}
\alias{pareto}
\title{Pareto function}
\usage{
pareto(x, mn = "Pareto barplot", ...)
}
\arguments{
\item{x}{A vector or factor containing the categorical data to be visualized.}

\item{mn}{An optional title for the chart (default is NULL).}

\item{...}{Additional arguments passed to the underlying plotting functions.}
}
\value{
A Pareto chart displaying the frequency of categories and their cumulative percentage.
}
\description{
Pareto function
}
\examples{
x <- c(10, 20, 30, 40, 50)
pareto(x)
}
