% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myread.R
\name{myread}
\alias{myread}
\title{Read a CSV file}
\usage{
myread(csv)
}
\arguments{
\item{csv}{A string specifying the CSV filename.}
}
\value{
A data frame containing the contents of the CSV file.
}
\description{
Read a CSV file
}
\examples{
myread("data.csv")
}
